Fix TransactionInactiveError issue in Safari/webkit browser
https://www.w3.org/TR/2018/REC-IndexedDB-2-20180130/#transaction-lifetime-concept

Each IDBTransaction has a lifetime detemined by the event loop context:
The lifetime of a transaction ends when the stackstack which the
transaction creation call() belongs end.

Here is the spec for transaction() method

'''
The transaction(storeNames, mode) method, when invoked, must run these
steps:

If a running upgrade transaction is associated with the connection,
throw an "InvalidStateError" DOMException.

If the connection's close pending flag is set, throw an
"InvalidStateError" DOMException.

Let scope be the set of unique strings in storeNames if it is a
sequence, or a set containing one string equal to storeNames otherwise.

If any string in scope is not the name of an object store in the
connected database, throw a "NotFoundError" DOMException.

If scope is empty, throw an "InvalidAccessError" DOMException.

If mode is not "readonly" or "readwrite", throw a TypeError.

Let transaction be a newly created transaction with connection, mode and
the set of object stores named in scope.

**Set transaction’s cleanup event loop to the current event loop.**

Return an IDBTransaction object representing transaction.
'''